gamsoflask/requirements.txt
python-dotenv
itchio
pandas
flask